% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Crosstalk_analysis.R
\name{Aggregate_Causality}
\alias{Aggregate_Causality}
\title{Calculate the TRS values}
\usage{
Aggregate_Causality(CC_results, data_type = "Target")
}
\arguments{
\item{CC_results}{The dataframe of PRS values.}

\item{data_type}{The type of the KeyGene (SSC or target).}
}
\value{
The dataframe of TRS.
}
\description{
Calculate the TRS values
}
\examples{
\donttest{
CC_pair_results <- Aggregate_Causality(CC_results, data_type = "Target")
}
}
