% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{Calculate_Average_Expression}
\alias{Calculate_Average_Expression}
\title{Calculate the average expression of a gene}
\usage{
Calculate_Average_Expression(targene, Expmat, nonzero = T)
}
\arguments{
\item{targene}{The gene to be calculated}

\item{Expmat}{The expression matrix}

\item{nonzero}{Whether consider cells with 0 targene expression}
}
\value{
The "sum" of the data
}
\description{
Calculate the average expression of a gene
}
