% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Crosstalk_analysis.R
\name{Calculate_Fidelity_Matrix}
\alias{Calculate_Fidelity_Matrix}
\title{Obtain a FIDELITY matrix for a fixed target gene.}
\usage{
Calculate_Fidelity_Matrix(CC_results, KeyTG)
}
\arguments{
\item{CC_results}{The dataframe of PRS values.}

\item{KeyTG}{The target to be focused on.}
}
\value{
The list containing the fidelity matrix within 3 types crosstalk modules.
}
\description{
Obtain a FIDELITY matrix for a fixed target gene.
}
\examples{
\donttest{
Fid_mat <- Calculate_Fidelity_Matrix(CC_results, KeyTG)
}
}
