% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLNconstruct.R
\name{Calculate_Importance}
\alias{Calculate_Importance}
\title{Extracting the PRS values from the regression models.}
\usage{
Calculate_Importance(ModelList)
}
\arguments{
\item{ModelList}{The list of regression model objects.}
}
\value{
The dataframe of PRS values.
}
\description{
Extracting the PRS values from the regression models.
}
