% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Crosstalk_analysis.R
\name{Calculate_Pairwise}
\alias{Calculate_Pairwise}
\title{Calculate the pairwise pathway fidelity/specificity}
\usage{
Calculate_Pairwise(CC_pair_results, KeyGene = NULL, type = "Fid")
}
\arguments{
\item{CC_pair_results}{The dataframe of TRS values.}

\item{KeyGene}{The Signal/Target to be focused on}

\item{type}{Whether to calculate fidelity (Fid) or specificity (Spe)}
}
\value{
The array containing the fidelity of the signal, out of all the pathways with the same Signal and SSC.
}
\description{
Calculate the pairwise pathway fidelity/specificity
}
\examples{
\donttest{
df <- Calculate_Pairwise(CC_pair_results, KeyGene = NULL, type = "Fid")
}
}
