% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Crosstalk_analysis.R
\name{Calculate_Specificity_Matrix}
\alias{Calculate_Specificity_Matrix}
\title{Obtain a SPECIFICITY matrix for a fixed signal(receptor).}
\usage{
Calculate_Specificity_Matrix(CC_results, KeyRec)
}
\arguments{
\item{CC_results}{The dataframe of PRS values.}

\item{KeyRec}{The signal(receptor) to be focused on.}
}
\value{
The list containing the specificity matrix within 3 types crosstalk modules.
}
\description{
Obtain a SPECIFICITY matrix for a fixed signal(receptor).
}
\examples{
\donttest{
Spe_mat <- Calculate_Specificity_Matrix(CC_results, KeyRec)
}
}
