% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{Chisq_Test}
\alias{Chisq_Test}
\title{Run the Chi-square's test}
\usage{
Chisq_Test(subset1, subset2, background)
}
\arguments{
\item{subset1}{The first set of genes.}

\item{subset2}{The second set of genes.}

\item{background}{All the expressed genes}
}
\value{
The p-value of the test
}
\description{
Run the Chi-square's test
}
