% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Crosstalk_analysis.R
\name{Count_Crosstalk}
\alias{Count_Crosstalk}
\title{Count the crosstalk pathways for each gene.}
\usage{
Count_Crosstalk(CC_results, KeyGenes = NULL, data_type = "SSC", verbose = T)
}
\arguments{
\item{CC_results}{The dataframe of PRS values.}

\item{KeyGenes}{The genes to be considered. By default, will consider all the genes in the pathways}

\item{data_type}{The type of the KeyGene (SSC or Target).}

\item{verbose}{Whether show notifications}
}
\value{
The dataframe containing the number of crosstalk pathways of each gene.
}
\description{
Count the crosstalk pathways for each gene.
}
\examples{
\donttest{
counts <- Count_Crosstalk(CC_results, data_type = "Target")
}
}
