% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{Extract_LR_Prob}
\alias{Extract_LR_Prob}
\title{Convert from a cellchat object an Ligand-Receptor probability dataframe}
\usage{
Extract_LR_Prob(
  result,
  source_type = NULL,
  target_type = NULL,
  pv_threshold = 0.05
)
}
\arguments{
\item{result}{The CellChat object}

\item{source_type}{The sender cell type.}

\item{target_type}{The receiver cell type.}

\item{pv_threshold}{The p-value threshold for filtering out the insignificant LR pairs.}
}
\value{
The expression matrix of the certain cell type.
}
\description{
Convert from a cellchat object an Ligand-Receptor probability dataframe
}
