% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{Filter_DB}
\alias{Filter_DB}
\title{Filter the prior database using genes expressed in the dataset}
\usage{
Filter_DB(DB, allgenes)
}
\arguments{
\item{DB}{The raw prior database, whose first two columns contain the "from" and "to" genes, respectively}

\item{allgenes}{All the expressed genes in the dataset.}
}
\value{
The filtered database.
}
\description{
Filter the prior database using genes expressed in the dataset
}
