% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{Filter_results}
\alias{Filter_results}
\title{Filter the Signal-SSC-Target pathways with low PRS results}
\usage{
Filter_results(results, PRS_thres = 0.05, remove_genes = T)
}
\arguments{
\item{results}{The raw PRS results generated by PRS_calc() function}

\item{PRS_thres}{The threshold for the PRS value. Pathways with PRS lower than PRS_thres*max(PRS) will be removed}

\item{remove_genes}{Boolean variables on whether remove mitochondrial and ribosomal genes}
}
\value{
The filtered dataframe containing the PRS values of pathways
}
\description{
Filter the Signal-SSC-Target pathways with low PRS results
}
\examples{
\donttest{
RecTFTG_filtered <- Filter_results(ress, PRS_thres = 0.05)
}
}
