% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLNconstruct.R
\name{FindRegulator}
\alias{FindRegulator}
\title{Infer regulator and their corresponding TGs.}
\usage{
FindRegulator(DB, exp, selected_genes = NULL, pv_threshold = 1)
}
\arguments{
\item{DB}{The gene-gene interaction database, normally between TF and TG.}

\item{exp}{The expression matrix.}

\item{selected_genes}{Select specific TGs. If set to NULL, will by default use all TGs in the database.}

\item{pv_threshold}{Threshold for the p-value of the Fisher test. Set to >= 1 if without extra filtering.}
}
\value{
The filtered database
}
\description{
Infer regulator and their corresponding TGs.
}
