% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{Fisher_Test}
\alias{Fisher_Test}
\title{Run the fisher's exact test}
\usage{
Fisher_Test(subset1, subset2, background)
}
\arguments{
\item{subset1}{The first set of genes.}

\item{subset2}{The second set of genes.}

\item{background}{All the expressed genes}
}
\value{
The p-value of the fisher test
}
\description{
Run the fisher's exact test
}
