% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{Get_Exp_Clu}
\alias{Get_Exp_Clu}
\title{Get the expression matrix from a seurat object}
\usage{
Get_Exp_Clu(
  SeuratObj,
  clusterID = NULL,
  assay = "RNA",
  datatype = "data",
  cutoff = 0.05
)
}
\arguments{
\item{SeuratObj}{The Seurat object of the dataset.}

\item{clusterID}{The cell type used (i.e. target_type).}

\item{assay}{The assay of the Seurat Object to use. By default use the "RNA" assay.}

\item{datatype}{The datatype of the assay. By default use the "scale.data".}

\item{cutoff}{The threshold for filtering out the low-expression genes.}
}
\value{
The expression matrix of the certain cell type.
}
\description{
Get the expression matrix from a seurat object
}
