% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLNconstruct.R
\name{Infer_CCI}
\alias{Infer_CCI}
\title{Infer the Cell-cell communication pattern in a given dataset.}
\usage{
Infer_CCI(SeuratObj, use_spatial = F, db_use = "human", scale_factors = NULL)
}
\arguments{
\item{SeuratObj}{The Seurat object of the dataset.}

\item{use_spatial}{Whether use spatial information. Set to True only when using spatial transcriptomics data.}

\item{db_use}{The species of the object. Set to human or mouse.}

\item{scale_factors}{The spatial scaling factors of the dataset. Use only when use_spatial = True.}
}
\value{
The CellChat Object containing Cell-cell communication information.
}
\description{
Infer the Cell-cell communication pattern in a given dataset.
}
\examples{
\donttest{
Infer_CCI(SeuratObj, use_spatial = F, db_use = "human", scale_factors = NULL)
}
}
