% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{is_connected_to}
\alias{is_connected_to}
\title{Find all nodes reachable by the given node.}
\usage{
is_connected_to(r, gg)
}
\arguments{
\item{r}{The seed node.}

\item{gg}{The graph object.}
}
\value{
All the nodes that the seed node can reach via the graph.
}
\description{
Find all nodes reachable by the given node.
}
