% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{myPageRank}
\alias{myPageRank}
\title{Perform the pagerank algorithm}
\usage{
myPageRank(gg, regulators, targets)
}
\arguments{
\item{gg}{The graph object of the database}

\item{regulators}{The receptors, which are also the seed of the algorithm.}

\item{targets}{The destinations of the paths.}
}
\value{
The PPR values of the certain regulator-target pair.
}
\description{
Perform the pagerank algorithm
}
\references{
Identify, quantify and characterize cellular communication from single-cell RNA sequencing data with scSeqComm. Bioinformatics (2022)

Browaeys, R., Saelens, W. & Saeys, Y. NicheNet: modeling intercellular communication by linking ligands to target genes. Nature Methods (2019)
}
