% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{nodes_find}
\alias{nodes_find}
\title{Find the nodes that contains the certain gene}
\usage{
nodes_find(x, gg, nod)
}
\arguments{
\item{x}{The target gene.}

\item{gg}{The graph object.}

\item{nod}{The nodes.}
}
\value{
All the nodes that contain gene x.
}
\description{
Find the nodes that contains the certain gene
}
