% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{Nonlinear_Sum}
\alias{Nonlinear_Sum}
\title{Calculate the different types of "sum"}
\usage{
Nonlinear_Sum(x, type = "Copula")
}
\arguments{
\item{x}{A numerical array}

\item{type}{The type of "sum": normal sum, copula sum, probability sum}
}
\value{
The "sum" of the data
}
\description{
Calculate the different types of "sum"
}
