% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualization.R
\name{PlotXT_Alluvial}
\alias{PlotXT_Alluvial}
\title{The river plot showing how the regulation flows}
\usage{
PlotXT_Alluvial(CC_results, KeyTG, min_weight = 0.45)
}
\arguments{
\item{CC_results}{The dataframe of PRS values.}

\item{KeyTG}{The target genes to be visualized.}

\item{min_weight}{The lower bound for the PRS of the pathways to be visualized}
}
\value{
An alluvial diagram of crosstalk
}
\description{
The river plot showing how the regulation flows
}
