% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualization.R
\name{PlotXT_Chord}
\alias{PlotXT_Chord}
\title{The chord diagram of PRS/fidelity/specificity values for a given gene that is shared}
\usage{
PlotXT_Chord(mat, orders = NULL, edge_colors = NULL)
}
\arguments{
\item{mat}{The matrix of PRS/fidelity/specificity values. Direction: columns to rows}

\item{orders}{The order of the genes. By default, orders <- c(rownames(mat), colnames(mat))}

\item{edge_colors}{The colors of the edges.}
}
\value{
A chord diagram of PRS/fidelity/specificity values for a given gene that is shared.
}
\description{
The chord diagram of PRS/fidelity/specificity values for a given gene that is shared
}
