% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualization.R
\name{PlotXT_Counts}
\alias{PlotXT_Counts}
\title{Show how many xtalk modules in the dataset}
\usage{
PlotXT_Counts(
  CC_results,
  KeyGenes = NULL,
  data_type = "Target",
  top_percent = 10
)
}
\arguments{
\item{CC_results}{The dataframe of PRS values.}

\item{KeyGenes}{The genes to be considered. By default, will consider all the genes in the pathways}

\item{data_type}{The type of the KeyGene (SSC or target).}

\item{top_percent}{In the inner panel, will show the the genes with most crosstalk pathways}
}
\value{
Two diagrams. Outer: the distribution of number of crosstalk pathways; Inner: the top-k targets
}
\description{
Show how many xtalk modules in the dataset
}
