% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualization.R
\name{PlotXT_FidSpe}
\alias{PlotXT_FidSpe}
\title{The heatmap of fidelity and specificity}
\usage{
PlotXT_FidSpe(CC_results, KeyTG, threshold = 0.15)
}
\arguments{
\item{CC_results}{The dataframe of PRS values.}

\item{KeyTG}{The target gene to be visualized.}

\item{threshold}{The lower bound for the fidelity of the pathways to be visualized}
}
\value{
Two diagrams. Left: the fidelity matrix; Right: the specificity matrix.
}
\description{
The heatmap of fidelity and specificity
}
