% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualization.R
\name{PlotXT_HeatMap}
\alias{PlotXT_HeatMap}
\title{The heatmap of PRS values for a given gene that is shared}
\usage{
PlotXT_HeatMap(CC_results, gene_used, genetype, topk = 25)
}
\arguments{
\item{CC_results}{The dataframe of PRS values.}

\item{gene_used}{The gene that is shared by pathways.}

\item{genetype}{The type of the shared gene: Receptor, SSC or Target}

\item{topk}{The maximum number of genes to be plotted.}
}
\value{
A heatmap of the PRS or the normalized PRS.Side bars contain the distribution of fidelity or specificity values.
}
\description{
The heatmap of PRS values for a given gene that is shared
}
\references{
Gu, Z. Complex heatmaps reveal patterns and correlations in multidimensional genomic data. Bioinformatics 2016.
}
