% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualization.R
\name{PlotXT_MultiCircularBar}
\alias{PlotXT_MultiCircularBar}
\title{The circulized bar charts of the specificity of Top-k specific targets for different conditions or signals}
\usage{
PlotXT_MultiCircularBar(df, KeyFactors = NULL, topk = 5, label_max = NULL)
}
\arguments{
\item{df}{The specificity dataframe}

\item{KeyFactors}{The conditions or signals. By default will use the second column of df.}

\item{topk}{The maximum number of targets to be plotted.}

\item{label_max}{The maximum specificity values shown in the charts.}
}
\value{
A circulized bar charts of the specificity of Top-k specific targets for different conditions or signals.
}
\description{
The circulized bar charts of the specificity of Top-k specific targets for different conditions or signals
}
