% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualization.R
\name{PlotXT_RecTGHeatmap}
\alias{PlotXT_RecTGHeatmap}
\title{The heatmap showing the TRS of signal-target pairs}
\usage{
PlotXT_RecTGHeatmap(CC_pair_results, Exp_clu, KeyTG, topk = 25)
}
\arguments{
\item{CC_pair_results}{The dataframe of TRS values.}

\item{Exp_clu}{The expression matrix of the target type.}

\item{KeyTG}{The genes to be considered.}

\item{topk}{The number of signal-target pairs to be plotted.}
}
\value{
Two diagrams. Upper: the contribution of each signal on the expression of each target; Lower: the heatmap of the TRS values.
}
\description{
The heatmap showing the TRS of signal-target pairs
}
