% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualization.R
\name{PlotXT_Ridgeline}
\alias{PlotXT_Ridgeline}
\title{The ridgeline plot of the distribution of fidelity}
\usage{
PlotXT_Ridgeline(CC_results, KeyTG)
}
\arguments{
\item{CC_results}{The dataframe of PRS values.}

\item{KeyTG}{The target gene to be visualized.}
}
\value{
A ridgeline plot of the distribution of fidelity for the signals that regulate the target via all the possible SSCs.
}
\description{
The ridgeline plot of the distribution of fidelity
}
