% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLNconstruct.R
\name{Prepare_Input}
\alias{Prepare_Input}
\title{Prepare all the input to the HGNN module.}
\usage{
Prepare_Input(
  SeuratObj,
  target_type,
  TGs,
  CCC_results,
  RTFDB,
  TFTGDB,
  data_dir,
  assay = "RNA",
  datatype = "scale.data",
  exp_threshold = 0.1,
  CCC_threshold = 0.05,
  Fisher_threshold = 1,
  species = "human"
)
}
\arguments{
\item{SeuratObj}{The Seurat object of the dataset.}

\item{target_type}{The receiver cell type.}

\item{TGs}{The target genes used.}

\item{CCC_results}{The Cellchat object.}

\item{RTFDB}{The receptor-TF database.}

\item{TFTGDB}{The TF-TG database.}

\item{data_dir}{The output directory of this function, which is also the input directory of the HGNN+ module.}

\item{assay}{The assay of the Seurat Object to use. By default use the "RNA" assay.}

\item{datatype}{The datatype of the assay. By default use the "scale.data".}

\item{exp_threshold}{The threshold for filtering out the low-expression genes.}

\item{CCC_threshold}{The threshold for filtering out lowly activated Ligand-Receptor pairs.}

\item{Fisher_threshold}{The threshold for the p-value of the Fisher test. Set to >= 1 if without extra filtering}

\item{species}{The species of the dataset. If set to "mouse" the gene names will be converted.}
}
\value{
The prepared databases, and the expression matrix that is stored in the data_dir for HGNN module.
}
\description{
Prepare all the input to the HGNN module.
}
\examples{
\donttest{
Prepare_Input <- function(SeuratObj, target_type, TGs, CCC_results, RTFDB, TFTGDB, data_dir)
}
}
