% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLNconstruct.R
\name{PRS_calc}
\alias{PRS_calc}
\title{Calculate the PRS values.}
\usage{
PRS_calc(Exp_clu, RTT_results, cutoff = 0.15)
}
\arguments{
\item{Exp_clu}{The expression matrix.}

\item{RTT_results}{The dataframe of activated pathways, generated from the HGNN module.}

\item{cutoff}{The threshold for filtering out lowly co-expresssed TF-TG pairs.}
}
\value{
The dataframe of PRS values.
}
\description{
Calculate the PRS values.
}
\examples{
\donttest{
RecTFTG <- PRS_calc(Exp_clu, ress, cutoff = 0.15)
}
}
