% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Databases.R
\docType{data}
\name{RTF_mouse}
\alias{RTF_mouse}
\title{Mouse Receptor-TF interactions}
\format{
A dataframe containing receptors, TFs, sources and databases
}
\source{
\url{https://github.com/LithiumHou/SigXTalk}
}
\usage{
RTF_mouse
}
\description{
A database for the biological interactions from receptors to TFs in Mouse
}
\references{
Browaeys, R., Saelens, W. & Saeys, Y. NicheNet: modeling intercellular communication by linking ligands to target genes. Nature Methods (2019)
}
\keyword{datasets}
