% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{Run_py_script}
\alias{Run_py_script}
\title{Run a Python script with given conda environment and arguments}
\usage{
Run_py_script(script_path, conda_env, args = character())
}
\arguments{
\item{script_path}{Character. Path to the Python script.}

\item{conda_env}{Character. The name of the conda environment.}

\item{args}{Character vector. Arguments to pass to the script (e.g., \code{c("--arg1", "value1","--arg2", "value2")}).}
}
\value{
No return value.
}
\description{
Run a Python script with given conda environment and arguments
}
\examples{
\dontrun{
Run_py_script(
  script_path = "./main.py",
  conda_env = "SigXTalk_py",
  args = c("--project",args.project,"--target_type",target_type)
)
}
}
