% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLNconstruct.R
\name{Run_Regression}
\alias{Run_Regression}
\title{Run the single expression.}
\usage{
Run_Regression(data_use)
}
\arguments{
\item{data_use}{The expression of genes.}
}
\value{
The regression model object.
}
\description{
Run the single expression.
}
