% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{single_receptor_ppr_wrapper}
\alias{single_receptor_ppr_wrapper}
\title{Do single time PPR algorithm}
\usage{
single_receptor_ppr_wrapper(r, E, graph, delta)
}
\arguments{
\item{r}{The seed node.}

\item{E}{The edges of the graph.}

\item{graph}{The graph object.}

\item{delta}{The dampling factor.}
}
\value{
The PPR matrix.
}
\description{
Do single time PPR algorithm
}
