% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Databases.R
\docType{data}
\name{TFT_mouse}
\alias{TFT_mouse}
\title{Mouse TF-Target gene interactions}
\format{
A dataframe containing TFs, target genes, sources and databases
}
\source{
\url{https://github.com/LithiumHou/SigXTalk}
}
\usage{
TFT_mouse
}
\description{
A database for the biological interactions from TFs to Targets in Mouse
}
\references{
Browaeys, R., Saelens, W. & Saeys, Y. NicheNet: modeling intercellular communication by linking ligands to target genes. Nature Methods (2019)
}
\keyword{datasets}
