% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{Trimean}
\alias{Trimean}
\title{Calculate the average value of a vector}
\usage{
Trimean(x, nonzero = T)
}
\arguments{
\item{x}{The vector.}

\item{nonzero}{Whether filter out non-zero values in the vector.}
}
\value{
The tri-mean of the vector.
}
\description{
Calculate the average value of a vector
}
