% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualization.R
\name{PlotCCI_ChordPlot}
\alias{PlotCCI_ChordPlot}
\title{Visualize the LRI using the chord diagram}
\usage{
PlotCCI_ChordPlot(result, topk = 10)
}
\arguments{
\item{result}{The LR pair probability extracted from the Extract_LR_Prob() function}

\item{topk}{The number of LR pairs that are visualized}
}
\description{
Visualize the LRI using the chord diagram
}
